+++ src/main/java/com/dico/blogs/user/model/User.java
@@ -16,7 +16,7 @@ import java.util.Set;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long Id;
+    private Integer Id;
 
     private String firstname;
     private String lastname;
@@ -35,11 +35,11 @@ public class User {
     Set<Role> roles = new HashSet<>();
 
 
-    public Long getId() {
+    public Integer getId() {
         return Id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         Id = id;
     }
