+++ src/main/java/com/dico/blogs/user/repository/RoleRepository.java
@@ -13,6 +13,6 @@ public interface RoleRepository extends JpaRepository<Role, Integer> {
             value = "SELECT * FROM role WHERE id NOT IN (SELECT role_id FROM user_role WHERE user_id = ?1)",
             nativeQuery = true
     )
-    List<Role> getUserNotRoles(Long id);
+    List<Role> getUserNotRoles(Integer id);
 
 }
